% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Argument.R
\name{Argument}
\alias{Argument}
\title{Create an Argument class object}
\usage{
Argument(name, flag = NULL, choices = NULL, default = NULL,
  required = FALSE, nargs = 1, action = "", type = "character",
  help = character())
}
\arguments{
\item{name}{character name of the argument (required)}

\item{flag}{character the pattern used to search the argument list, by default name}

\item{choices}{a vector of the choices if limited to a listing}

\item{default}{the default value if none provided}

\item{required}{logical, set to TRUE is required otherwise it is optional}

\item{nargs}{the number of arguments (maximum)}

\item{action}{a function name or NULL}

\item{type}{character, one of the following \code{logical, character, integer, double,
numeric, set_true, set_false}  \code{set_true} and \code{set_false} are logicals 
based upon presence/abscence such as --plot or --no-plot  Alternatively,
use 'logical' for "-plot  TRUE" or "-plot  FALSE" style booleans.}

\item{help}{character vector of helpful hints}
}
\value{
an ArgumentRefClass object
}
\description{
Create an Argument class object
}
\seealso{
Other Argument: \code{\link{ArgumentRefClass_getAction}},
  \code{\link{ArgumentRefClass_get_usage}},
  \code{\link{ArgumentRefClass_get}},
  \code{\link{ArgumentRefClass_parse_argument}},
  \code{\link{ArgumentRefClass_print_help}},
  \code{\link{ArgumentRefClass_show}},
  \code{\link{parseArgument}}
}

