% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Argument.R
\docType{class}
\name{ArgumentRefClass-class}
\alias{ArgumentRefClass}
\alias{ArgumentRefClass-class}
\title{ArgumentRefClass to manage one argument}
\description{
ArgumentRefClass to manage one argument
}
\section{Fields}{

\describe{
\item{\code{name}}{the name of the argument - typically the name provided in the argument string}

\item{\code{flag}}{the characater string used to search for the argument.  By default
the name, prepended by \code{--} or \code{-}.  Allowing name and flag to be different
is to allow the user to specify a name like 'bosworth' for an argumeent flagged as '-b'}

\item{\code{value}}{the parsed argument value as a list.}

\item{\code{choices}}{a list of allowable argument values (like having a dropdown list)}

\item{\code{nargs}}{the number of expected arguments, by default 1}

\item{\code{required}}{logical - must the user provide a value?}

\item{\code{type}}{charcater specifies the type of argument.  Allowed values are
\describe{
   \item{logical}{Explicitly set TRUE/FALSE arguments}
   \item{set_true}{For arguments set by presence ala --plot}
   \item{set_false}{For argument set by presence ala --no-plot}
   \item{character}{For character arguments}
   \item{numeric}{For numeric arguments ala integers and doubles}
   \item{integer}{For explictly integer numerics}
   \item{double}{For explictly double precision numerics}
}}

\item{\code{action}}{the name of a function to call with the argument}

\item{\code{default}}{the optional default value}

\item{\code{help}}{a character vector of help text}
}}

