% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Argument.R
\name{ArgumentRefClass_get}
\alias{ArgumentRefClass_get}
\title{Get the value(s) of an argument}
\arguments{
\item{...}{unspecified arguments for the Argument$getAction() method}
}
\value{
the value of the argument (type varies)
}
\description{
Get the value(s) of an argument
}
\seealso{
Other Argument: \code{\link{ArgumentRefClass_getAction}};
  \code{\link{ArgumentRefClass_get_usage}};
  \code{\link{ArgumentRefClass_parse_argument}};
  \code{\link{ArgumentRefClass_print_help}};
  \code{\link{ArgumentRefClass_show}};
  \code{\link{Argument}}; \code{\link{parseArgument}}
}

