% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Argument.R
\name{ArgumentRefClass_getAction}
\alias{ArgumentRefClass_getAction}
\title{Run the 'action' function}
\arguments{
\item{...}{unspecified arguments}
}
\value{
the value of the argument (type varies)
}
\description{
Run the 'action' function
}
\seealso{
Other Argument: \code{\link{ArgumentRefClass_get_usage}},
  \code{\link{ArgumentRefClass_get}},
  \code{\link{ArgumentRefClass_parse_argument}},
  \code{\link{ArgumentRefClass_print_help}},
  \code{\link{ArgumentRefClass_show}},
  \code{\link{Argument}}, \code{\link{parseArgument}}
}
