% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Argument.R
\name{ArgumentRefClass_parse_argument}
\alias{ArgumentRefClass_parse_argument}
\title{Parse the argument list selecting the one of interest}
\arguments{
\item{x}{character argument list}
}
\value{
logical indicating success
}
\description{
Parse the argument list selecting the one of interest
}
\seealso{
Other Argument: \code{\link{ArgumentRefClass_getAction}};
  \code{\link{ArgumentRefClass_get_usage}};
  \code{\link{ArgumentRefClass_get}};
  \code{\link{ArgumentRefClass_print_help}};
  \code{\link{ArgumentRefClass_show}};
  \code{\link{Argument}}; \code{\link{parseArgument}}
}

