% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mail.R
\name{charliemail}
\alias{charliemail}
\title{Send a simple mail via mail, but hardwired for charlie}
\usage{
charliemail(to = "btupper@bigelow.org", sbj = "charliemail",
  msg = paste("at", Sys.time(), "you have mail mail"), attachment = NA,
  verbose = TRUE)
}
\arguments{
\item{to}{a character vector of one or more valid email addresses}

\item{sbj}{a character for the subject line (required)}

\item{msg}{- a character vector of one or more lines for the message body (NA to skip)}

\item{attachment}{the fully qualified filename to attach (if any, NA to skip)}

\item{verbose}{logical, if TRUE then echo the command}
}
\value{
0 for success and non-zero otherwise
}
\description{
Send a simple mail via mail, but hardwired for charlie
}
\seealso{
Other mail: \code{\link{has_mail}}, \code{\link{has_mutt}},
  \code{\link{has_nail}}, \code{\link{mailmail}},
  \code{\link{muttmail}}, \code{\link{nailmail}},
  \code{\link{sendmail}}
}
