% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CommandArgs.R
\name{CommandArgs}
\alias{CommandArgs}
\title{Generate a CommandArgs reference}
\usage{
CommandArgs(args = commandArgs(trailingOnly = FALSE), name = "program_name",
  help = NULL)
}
\arguments{
\item{args}{a character vector as returned by \code{\link{commandArgs}}
or NULL}

\item{help}{a character vector of helpful information}
}
\value{
a CommandArgsRefClass instance
}
\description{
Generate a CommandArgs reference
}
\seealso{
Other CommandArgs: \code{\link{CommandArgs_add_argument}};
  \code{\link{CommandArgs_add}};
  \code{\link{CommandArgs_get_all}};
  \code{\link{CommandArgs_get}},
  \code{\link{CommandArgs_get}};
  \code{\link{CommandArgs_help_called}};
  \code{\link{CommandArgs_parse_arguments}};
  \code{\link{CommandArgs_print_help}};
  \code{\link{CommandArgs_show}}
}

