% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommandArgs.R
\name{CommandArgs_add_argument}
\alias{CommandArgs_add_argument}
\title{Add an argument definition
See \code{link{Argument}} for input details}
\arguments{
\item{name}{the name of the argument to add}

\item{...}{furtehr arguments for the \code{\link{Argument}} to be defined}
}
\description{
Add an argument definition
See \code{link{Argument}} for input details
}
\seealso{
Other CommandArgs: \code{\link{CommandArgs_add}},
  \code{\link{CommandArgs_get_all}},
  \code{\link{CommandArgs_get}},
  \code{\link{CommandArgs_help_called}},
  \code{\link{CommandArgs_parse_arguments}},
  \code{\link{CommandArgs_print_help}},
  \code{\link{CommandArgs_show}}, \code{\link{CommandArgs}}
}
\concept{CommandArgs}
