% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CommandArgs.R
\name{CommandArgs_get}
\alias{CommandArgs_get}
\title{Test is the named argument is present}
\arguments{
\item{name}{the name of the argument to test}

\item{name}{the name of the argument to retrieve}

\item{...}{futher arguments for the \code{Argument$get()} method requested}
}
\value{
logical indicating is (or is not) present

the value of the argument (type varies)
}
\description{
Test is the named argument is present

Get the value(s) of an argument
}
\seealso{
Other CommandArgs: \code{\link{CommandArgs_add_argument}};
  \code{\link{CommandArgs_add}};
  \code{\link{CommandArgs_get_all}};
  \code{\link{CommandArgs_parse_arguments}};
  \code{\link{CommandArgs_print_help}};
  \code{\link{CommandArgs_show}}; \code{\link{CommandArgs}}

Other CommandArgs: \code{\link{CommandArgs_add_argument}};
  \code{\link{CommandArgs_add}};
  \code{\link{CommandArgs_get_all}};
  \code{\link{CommandArgs_parse_arguments}};
  \code{\link{CommandArgs_print_help}};
  \code{\link{CommandArgs_show}}; \code{\link{CommandArgs}}
}

