% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CommandArgs.R
\name{CommandArgs_get_all}
\alias{CommandArgs_get_all}
\title{Get all arguments}
\arguments{
\item{...}{futher arguments for the \code{Argument$get()} method requested}
}
\value{
a list of key-value pairs
}
\description{
Get all arguments
}
\seealso{
Other CommandArgs: \code{\link{CommandArgs_add_argument}};
  \code{\link{CommandArgs_add}};
  \code{\link{CommandArgs_get}},
  \code{\link{CommandArgs_get}};
  \code{\link{CommandArgs_parse_arguments}};
  \code{\link{CommandArgs_print_help}};
  \code{\link{CommandArgs_show}}; \code{\link{CommandArgs}}
}

