% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommandArgs.R
\name{CommandArgs_help_called}
\alias{CommandArgs_help_called}
\title{Detect if user called for help using "--help" or "-h"}
\value{
logical, TRUE if --help or -h exists
}
\description{
Detect if user called for help using "--help" or "-h"
}
\seealso{
Other CommandArgs: \code{\link{CommandArgs_add_argument}},
  \code{\link{CommandArgs_add}},
  \code{\link{CommandArgs_get_all}},
  \code{\link{CommandArgs_get}},
  \code{\link{CommandArgs_parse_arguments}},
  \code{\link{CommandArgs_print_help}},
  \code{\link{CommandArgs_show}}, \code{\link{CommandArgs}}
}

