% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Configurator.R
\name{Configurator}
\alias{Configurator}
\title{Create a new ConfiguratorRefClass and possibly read the contents of a 
config file.}
\usage{
Configurator(filename, sep = "=", comm = "#")
}
\arguments{
\item{filename}{a fully qualified path to a config file}

\item{sep}{character, the character separating key-value pairs, by default "="}

\item{comm}{character, the character found in the first spot of a comment line, by default "#"}
}
\value{
an instance of ConfiguratorRefClass
}
\description{
Create a new ConfiguratorRefClass and possibly read the contents of a 
config file.
}
\seealso{
Other CONFIGURATOR: \code{\link{ConfiguratorRefClass_get}},
  \code{\link{ConfiguratorRefClass_parse_line}},
  \code{\link{ConfiguratorRefClass_read_file}},
  \code{\link{ConfiguratorRefClass_set}},
  \code{\link{ConfiguratorRefClass_show}},
  \code{\link{ConfiguratorRefClass_write_file}}
}
\concept{CONFIGURATOR}
