% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Configurator.R
\docType{class}
\name{ConfiguratorRefClass-class}
\alias{ConfiguratorRefClass}
\alias{ConfiguratorRefClass-class}
\title{A class for configuration elements}
\description{
A class for configuration elements
}
\section{Fields}{

\describe{
\item{\code{filename}}{the fully qualified name of the configuration file}

\item{\code{container}}{a list of configuration elements - one element per named [section]}

\item{\code{comments}}{a list of configuration comments by [section] plus a [.header]}

\item{\code{sep}}{the default key-value pair separator ("=") for writing}

\item{\code{comm}}{the default comment character ("#") flagging comment lines}
}}
\section{Methods}{

\describe{
\item{\code{get(section, name, default = NULL)}}{Retrieve the specified item in the sectioned named or NULL of not found}

\item{\code{parse_section(s)}}{Given one or more raw text lines, parse the key=value pairs into the 
expected named character vector}

\item{\code{read_file(filename = .self$filename)}}{read the specified configuration file}

\item{\code{set(section, name, value)}}{Set the value of the specified item in the sectioned named}

\item{\code{write_file(filename = .self$filename, append = FALSE)}}{write the specified configuration file}
}}

