% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Configurator.R
\name{ConfiguratorRefClass_get}
\alias{ConfiguratorRefClass_get}
\title{Retrieve one keyword/value pair by section or an entire section}
\arguments{
\item{section}{the name of the [section] to search}

\item{name}{the keyword name of the value to retrieve, if missing then the contents of the section are returned}

\item{default}{the value to return if the key or section is not found (by default NULL)}
}
\value{
character value(s) or \code{default} if not found
}
\description{
Retrieve one keyword/value pair by section or an entire section
}
\examples{
\dontrun{
   X$get("mySection", "myName")
}
}
\seealso{
Other CONFIGURATOR: \code{\link{ConfiguratorRefClass_parse_line}};
  \code{\link{ConfiguratorRefClass_read_file}};
  \code{\link{ConfiguratorRefClass_set}};
  \code{\link{ConfiguratorRefClass_show}};
  \code{\link{ConfiguratorRefClass_write_file}};
  \code{\link{Configurator}}
}

