% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Configurator.R
\name{ConfiguratorRefClass_parse_line}
\alias{ConfiguratorRefClass_parse_line}
\title{Parse one or more raw text lines of key=value or key:value form.}
\arguments{
\item{a}{character vector of key=value sets}
}
\value{
a named character vector, names are keys data are the values
}
\description{
Parse one or more raw text lines of key=value or key:value form.
}
\seealso{
Other CONFIGURATOR: \code{\link{ConfiguratorRefClass_get}},
  \code{\link{ConfiguratorRefClass_read_file}},
  \code{\link{ConfiguratorRefClass_set}},
  \code{\link{ConfiguratorRefClass_show}},
  \code{\link{ConfiguratorRefClass_write_file}},
  \code{\link{Configurator}}
}
