% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Configurator.R
\name{ConfiguratorRefClass_read_file}
\alias{ConfiguratorRefClass_read_file}
\title{Read a config file comprised of [sections] of key=value pairs.}
\arguments{
\item{filename}{the fully qualified name of the configuration file}
}
\value{
the configurations stored in a list
}
\description{
Read a config file comprised of [sections] of key=value pairs.
}
\seealso{
Other CONFIGURATOR: \code{\link{ConfiguratorRefClass_get}},
  \code{\link{ConfiguratorRefClass_parse_line}},
  \code{\link{ConfiguratorRefClass_set}},
  \code{\link{ConfiguratorRefClass_show}},
  \code{\link{ConfiguratorRefClass_write_file}},
  \code{\link{Configurator}}
}

