% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Configurator.R
\name{ConfiguratorRefClass_set}
\alias{ConfiguratorRefClass_set}
\title{Set one keyword/value pair by section}
\arguments{
\item{section}{the name of the [section] to search}

\item{name}{the keyword name of the value to set}

\item{value}{the value, must survive \code{\link{as.character}} conversion}
}
\value{
logical TRUE if successful
}
\description{
Set one keyword/value pair by section
}
\examples{
\dontrun{
   X$set("mySection", "myName", someValue)
}
}
\seealso{
Other CONFIGURATOR: \code{\link{ConfiguratorRefClass_get}},
  \code{\link{ConfiguratorRefClass_parse_line}},
  \code{\link{ConfiguratorRefClass_read_file}},
  \code{\link{ConfiguratorRefClass_show}},
  \code{\link{ConfiguratorRefClass_write_file}},
  \code{\link{Configurator}}
}
