% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Configurator.R
\name{ConfiguratorRefClass_write_file}
\alias{ConfiguratorRefClass_write_file}
\title{Write a config file comprised of [sections] of key=value pairs.}
\arguments{
\item{filename}{the fully qualified name of the configuration file}

\item{append}{logical, if TRUE append to the specified file otherwise truncate}
}
\value{
logical named result form file.exists
}
\description{
Write a config file comprised of [sections] of key=value pairs.
}
\seealso{
Other CONFIGURATOR: \code{\link{ConfiguratorRefClass_get}};
  \code{\link{ConfiguratorRefClass_parse_line}};
  \code{\link{ConfiguratorRefClass_read_file}};
  \code{\link{ConfiguratorRefClass_set}};
  \code{\link{ConfiguratorRefClass_show}};
  \code{\link{Configurator}}
}

