% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{cp_dir}
\alias{cp_dir}
\title{Copy the contents of one directory to another using system's \code{cp}.}
\usage{
cp_dir(src, dst, flags = "-uR")
}
\arguments{
\item{src}{character the fully qualified source directory}

\item{dst}{character the fully qualified destination directory}

\item{flags}{charcater flags for cp, by default "-uR"}
}
\value{
the output of the systems cp command (0 = good!)
}
\description{
Note that differs from rsync_dirs
}
\seealso{
Other FILE: \code{\link{file_parts}},
  \code{\link{get_temp_dirname}},
  \code{\link{get_temp_filename}}, \code{\link{rsync_dirs}}
}

