% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommandArgs.R
\name{dummy_commandArgs}
\alias{dummy_commandArgs}
\title{Retrieve dummy arguments ala \code{commandArgs(trailingOnly = FALSE)}}
\usage{
dummy_commandArgs(os = c("linux", "darwin")[1])
}
\arguments{
\item{os}{character the name of the operating system to mimic.}
}
\value{
character vector of typical command arguments
}
\description{
Note that the value returned is a hint or reminder.  Your best bet is to test drive
your own configuration.
}
