% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{get_temp_dirname}
\alias{get_temp_dirname}
\title{Retrieve a temporary directory name}
\usage{
get_temp_dirname(pattern = "", ramdisk = file.exists("/dev/shm"),
  create = TRUE)
}
\arguments{
\item{pattern}{character if provided this will be prepended on the temp directory returned}

\item{ramdisk}{logical if TRUE, then try to get a path into the ramdisk first, if not
available then use the standard tempdir}

\item{create}{logical if TRUE then create the path}
}
\value{
the fully qualified temporary path
}
\description{
This fucntion attempts to create uniquely name directories.
}
\seealso{
Other FILE: \code{\link{cp_dir}}, \code{\link{file_parts}},
  \code{\link{get_temp_filename}}, \code{\link{rsync_dirs}}
}
\concept{FILE}
