% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{get_temp_filename}
\alias{get_temp_filename}
\title{Create a temporary file name}
\usage{
get_temp_filename(pattern = "", path = get_temp_dirname(), ext = "")
}
\arguments{
\item{pattern}{character if provided this will be prepended on the filename returned}

\item{path}{charcater the path to attach the file within}

\item{ext}{charcater any extension to add to the filename generated}
}
\description{
Create a temporary file name
}
\seealso{
Other FILE: \code{\link{cp_dir}}, \code{\link{file_parts}},
  \code{\link{get_temp_dirname}}, \code{\link{rsync_dirs}}
}
