% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mail.R
\name{has_mail}
\alias{has_mail}
\title{Test if mail is installed on the host platform}
\usage{
has_mail()
}
\value{
logical, TRUE if present
}
\description{
Test if mail is installed on the host platform
}
\seealso{
Other mail: \code{\link{has_mutt}}, \code{\link{has_nail}},
  \code{\link{mailmail}}, \code{\link{muttmail}},
  \code{\link{nailmail}}, \code{\link{sendmail}}
}
