% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mail.R
\name{has_nail}
\alias{has_nail}
\title{Test if nail is installed on the host platform
See \url{http://nail.sourceforge.net/}}
\usage{
has_nail()
}
\value{
logical, TRUE if present
}
\description{
Test if nail is installed on the host platform
See \url{http://nail.sourceforge.net/}
}
\seealso{
Other mail: \code{\link{has_mutt}}, \code{\link{muttmail}},
  \code{\link{nailmail}}, \code{\link{sendmail}}
}
\concept{mail}
