% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/compressed.R
\name{is_tgz}
\alias{is_tgz}
\title{Test if a file is tarred and gzip (.tar, .gz or tgz)}
\usage{
is_tgz(filename)
}
\arguments{
\item{filename}{the fully qualified name of the file to test}
}
\value{
logical, TRUE if the input is likely gzipped
}
\description{
We guess that the first four bytes tell us enough to know what kind of zip
file we have.  For more information see \url{http://pank.org/blog/archives/000202.html}
tgz's first four bytes will be 1f 8b 08 00
}
\seealso{
Other COMPRESSED: \code{\link{is_bzip2}};
  \code{\link{is_gzip}}; \code{\link{is_zip}};
  \code{\link{pack_directory}};
  \code{\link{test_compressed}}; \code{\link{un_bzip2}};
  \code{\link{un_gz}}; \code{\link{un_tgz}};
  \code{\link{un_zip}}; \code{\link{unpack_archive}}
}

