% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compressed.R
\name{is_zip}
\alias{is_zip}
\title{Test if a file is zipped.}
\usage{
is_zip(filename)
}
\arguments{
\item{filename}{the fully qualified name of the file to test}
}
\value{
logical, TRUE if the input is likely zipped
}
\description{
Test if a file is zipped.
}
\seealso{
Other COMPRESSED: \code{\link{is_bzip2}},
  \code{\link{is_gzip}}, \code{\link{is_tgz}},
  \code{\link{pack_directory}},
  \code{\link{test_compressed}}, \code{\link{un_bzip2}},
  \code{\link{un_gz}}, \code{\link{un_tgz}},
  \code{\link{un_zip}}, \code{\link{unpack_archive}}
}
