% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Logger.R
\name{Logger}
\alias{Logger}
\title{Create a LoggerRefClass object}
\usage{
Logger(name = "logger", filename = NA, do_echo = TRUE,
  do_timestamp = TRUE, start_watch = TRUE)
}
\arguments{
\item{name}{the name of the logger, I usually make it the name of the script}

\item{filename}{the name of the file to send the messages to}

\item{do_echo}{logical to output to sterr}

\item{do_timestamp}{if TRUE then prepend a time stamp to messages}

\item{start_watch}{logical, if TRUE start the stopwatch}
}
\value{
a LoggerRefClass object
}
\description{
Create a LoggerRefClass object
}
\seealso{
Other Logger: \code{\link{Logger_elapsed}};
  \code{\link{Logger_error}}; \code{\link{Logger_info}};
  \code{\link{Logger_log}};
  \code{\link{Logger_scriptreport}};
  \code{\link{Logger_setfilename}};
  \code{\link{Logger_warn}}
}

