% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Logger.R
\name{Logger_elapsed}
\alias{Logger_elapsed}
\title{Output the elapsed time}
\description{
Useage: X$elapsed(name = 1)
}
\seealso{
Other Logger: \code{\link{Logger_error}},
  \code{\link{Logger_info}}, \code{\link{Logger_log}},
  \code{\link{Logger_scriptreport}},
  \code{\link{Logger_setfilename}},
  \code{\link{Logger_warn}}, \code{\link{Logger}}
}
