% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Logger.R
\name{Logger_error}
\alias{Logger_error}
\title{Output an error message}
\arguments{
\item{...}{message parts, will be pasted together with paste.sep}

\item{leader}{character, statement prepended to the message}
}
\description{
Output an error message
}
\seealso{
Other Logger: \code{\link{Logger_elapsed}},
  \code{\link{Logger_info}}, \code{\link{Logger_log}},
  \code{\link{Logger_scriptreport}},
  \code{\link{Logger_setfilename}},
  \code{\link{Logger_warn}}, \code{\link{Logger}}
}
