% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Logger.R
\name{Logger_roll}
\alias{Logger_roll}
\title{Roll the log file to a new one based upon file size}
\arguments{
\item{max_size}{numeric the maximum log file size in bytes to trigger rotation}
}
\value{
invisible, the size of the current log file (possibly rotated)
}
\description{
Roll the log file to a new one based upon file size
}
