% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Logger.R
\name{Logger_scriptreport}
\alias{Logger_scriptreport}
\title{Create a script report}
\arguments{
\item{name}{character, the name of the script or what have you}

\item{args}{character vector of script arguments}

\item{do_echo}{logical, set to TRUE to echo to the command line}
}
\description{
Create a script report
}
\examples{
\dontrun{
   args <- commandArgs(trailingOnly = FALSE)
   X <- Logger(name = 'test log')
   X$scriptreport(args)
}
}
\seealso{
Other Logger: \code{\link{Logger_elapsed}};
  \code{\link{Logger_error}}; \code{\link{Logger_info}};
  \code{\link{Logger_log}};
  \code{\link{Logger_setfilename}};
  \code{\link{Logger_warn}}; \code{\link{Logger}}
}

