% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Logger.R
\name{Logger_setfilename}
\alias{Logger_setfilename}
\title{Set the filename attribute}
\arguments{
\item{file}{the name of the file}
}
\description{
Set the filename attribute
}
\seealso{
Other Logger: \code{\link{Logger_elapsed}},
  \code{\link{Logger_error}}, \code{\link{Logger_info}},
  \code{\link{Logger_log}},
  \code{\link{Logger_scriptreport}},
  \code{\link{Logger_tail}}, \code{\link{Logger_warn}},
  \code{\link{Logger}}
}
\concept{Logger}
