% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Logger.R
\name{Logger_tail}
\alias{Logger_tail}
\title{Retrieve the tail of a log file}
\arguments{
\item{n}{number of lines to retrieve}
}
\description{
Retrieve the tail of a log file
}
\seealso{
Other Logger: \code{\link{Logger_elapsed}},
  \code{\link{Logger_error}}, \code{\link{Logger_info}},
  \code{\link{Logger_log}},
  \code{\link{Logger_scriptreport}},
  \code{\link{Logger_setfilename}},
  \code{\link{Logger_warn}}, \code{\link{Logger}}
}
\concept{Logger}
