% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Logger.R
\name{Logger_warn}
\alias{Logger_warn}
\title{Output a warning message}
\arguments{
\item{...}{message parts, will be pasted together with paste.sep}

\item{leader}{character, statement prepended to the message}
}
\description{
Output a warning message
}
\seealso{
Other Logger: \code{\link{Logger_elapsed}},
  \code{\link{Logger_error}}, \code{\link{Logger_info}},
  \code{\link{Logger_log}},
  \code{\link{Logger_scriptreport}},
  \code{\link{Logger_setfilename}},
  \code{\link{Logger_tail}}, \code{\link{Logger}}
}
