% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Logger.R
\docType{class}
\name{LoggerRefClass-class}
\alias{LoggerRefClass}
\alias{LoggerRefClass-class}
\title{A Logger reference class for simple message logging}
\description{
A Logger reference class for simple message logging
}
\section{Fields}{

\describe{
\item{\code{name}}{character, the name of the Logger}

\item{\code{filename}}{the name of the file to write to}

\item{\code{timer}}{TimekeeperRefClass object}

\item{\code{.echo}}{logical, logical for internal use}

\item{\code{.hasfile}}{logical, for internal use}

\item{\code{.timestamp}}{logical, for internal use}
}}

