% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mail.R
\name{nailmail}
\alias{nailmail}
\title{Send a simple mail via nail  - see \url{http://linux.die.net/man/1/nail}}
\usage{
nailmail(to = "name@somewhere.org", sbj = "nail mail",
  msg = paste("at", Sys.time(), "you have nail mail"), attachment = NA,
  verbose = TRUE)
}
\arguments{
\item{to}{a character vector of one or more valid email addresses}

\item{sbj}{a character for the subject line (required)}

\item{msg}{- a character vector of one or more lines for the message body (NA to skip)}

\item{attachment}{the fully qualified filename to attach (if any, NA to skip)}

\item{verbose}{logical, if TRUE then echo the command}
}
\value{
0 for success and non-zero otherwise
}
\description{
Send a simple mail via nail  - see \url{http://linux.die.net/man/1/nail}
}
\seealso{
Other mail: \code{\link{has_mutt}}, \code{\link{has_nail}},
  \code{\link{muttmail}}, \code{\link{sendmail}}
}
\concept{mail}
