% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compressed.R
\name{pack_directory}
\alias{pack_directory}
\title{Create a zipped or tarred and gzipped archive of a directory}
\usage{
pack_directory(directory, fmt = c("tgz", "zip")[1],
  outfile = file.path(dirname(directory), paste0(basename(directory), c(tgz =
  ".tar.gz", zip = ".zip")[fmt])))
}
\arguments{
\item{directory}{one or more directories to pack. NOTE they are not packed together.}

\item{fmt}{character, indicates output format.  Currently just 'tgz' and 'zip'}

\item{outfile}{the full qualified name(s) of the output archive(s)}
}
\value{
named logical indicating the outfile exists (or not!)
}
\description{
Create a zipped or tarred and gzipped archive of a directory
}
\seealso{
Other COMPRESSED: \code{\link{is_bzip2}},
  \code{\link{is_gzip}}, \code{\link{is_tgz}},
  \code{\link{is_zip}}, \code{\link{test_compressed}},
  \code{\link{un_bzip2}}, \code{\link{un_gz}},
  \code{\link{un_tgz}}, \code{\link{un_zip}},
  \code{\link{unpack_archive}}
}

