% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Argument.R
\name{parseArgument}
\alias{parseArgument}
\title{Parse the argument character for this Argument's value(s)}
\usage{
parseArgument(X, x)
}
\arguments{
\item{X}{the ArgumentRefClass object}

\item{x}{the argument character vector to mine for flagged value}
}
\value{
logical indicating success/failure
}
\description{
Parse the argument character for this Argument's value(s)
}
\seealso{
Other Argument: \code{\link{ArgumentRefClass_getAction}},
  \code{\link{ArgumentRefClass_get_usage}},
  \code{\link{ArgumentRefClass_get}},
  \code{\link{ArgumentRefClass_parse_argument}},
  \code{\link{ArgumentRefClass_print_help}},
  \code{\link{ArgumentRefClass_show}},
  \code{\link{Argument}}
}
\concept{Argument}
