% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{rsync_dirs}
\alias{rsync_dirs}
\title{Sync two directories using the system \code{rsync}}
\usage{
rsync_dirs(src, dst, flags = "-rcv")
}
\arguments{
\item{src}{character the fully qualified source directory}

\item{dst}{character the fully qualified destination directory}

\item{flags}{charcater flags for rsync, by default '-rvc'}
}
\value{
the output of the systems rsync command (0 = good!)
}
\description{
In rsync there is a complete test on the change of a file before copying.  Using 
cp_dir perfoms a simpler test for changes, so it might be faster.  I think for
most pipeline uses it may be better to use cp_dir since pipelines are a once-and-done
task.
}
\seealso{
Other FILE: \code{\link{cp_dir}}, \code{\link{file_parts}},
  \code{\link{get_temp_dirname}},
  \code{\link{get_temp_filename}}
}
