% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mail.R
\name{sendmail}
\alias{sendmail}
\title{Send an email from R using the underlying mail services.}
\usage{
sendmail(..., what = c("mail", "nail", "mutt", "charlie")[1])
}
\arguments{
\item{...}{arguments for the mailer of choice. Currently, only mutt is supported.}

\item{what}{the name of the mail engine to use, currently 'nail' and 'mutt' are supported}
}
\value{
value returned by the system-level mail program, non-zero for failure.
}
\description{
A simple wrapper around the platforms email engine.  Currently \code{mail}, \code{mutt} and
\code{nail} are supported, but the host must have these installed.
Also supported is \code{charlie} whihc is a hardwired version of mail for use on charlie
}
\seealso{
Other mail: \code{\link{charliemail}},
  \code{\link{has_mail}}, \code{\link{has_mutt}},
  \code{\link{has_nail}}, \code{\link{mailmail}},
  \code{\link{muttmail}}, \code{\link{nailmail}}
}
