% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stopwatch.R
\name{Stopwatch}
\alias{Stopwatch}
\title{Create a StopwatchRefClass object}
\usage{
Stopwatch(name = "", start_watch = FALSE)
}
\arguments{
\item{name}{the name to assign to this object}

\item{start_watch}{logical, if TRUE then start the Stopwatch}
}
\value{
a \code{StopwatchRefClass} object
}
\description{
Create a StopwatchRefClass object
}
\examples{
\dontrun{
   # create stop watch
   X <- Stopwatch(name = 'timex', start = FALSE)
   #start it
   X$start()
   X$elapsed()
   X
   Sys.sleep(5)
   # When stopping you can name the interval for later retreival
   X$stop(name = 'first')
   X
   X$elapsed(name = "first")
   X$start()
   Sys.sleep(2)
   X$stop(name = "second")
   X$elapsed(name = 'second')
   X$elapsed(name = 'first')
   # now accumulate the elapsed times
   X$elapsed(accumulate = TRUE)}
}
\seealso{
Other Stopwatch Timekeeper: \code{\link{Timekeeper_elapsed}},
  \code{\link{Timekeeper_has_watch}},
  \code{\link{Timekeeper_reset}},
  \code{\link{Timekeeper_start}},
  \code{\link{Timekeeper_stop}}, \code{\link{Timekeeper}}
}
\concept{Stopwatch Timekeeper}
