% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Stopwatch.R
\name{Stopwatch_add_elapsed}
\alias{Stopwatch_add_elapsed}
\title{Add an elapsed time.  If the watch is running then
it is current less start time.  Users should not use this method directly.}
\arguments{
\item{interval}{elapsed time segments can be tagged by name for later retrieval by name}
}
\value{
the number of dt elements (invisibly)
}
\description{
Add an elapsed time.  If the watch is running then
it is current less start time.  Users should not use this method directly.
}
\seealso{
Other Stopwatch: \code{\link{Stopwatch_current_elapsed}};
  \code{\link{Stopwatch_elapsed}};
  \code{\link{Stopwatch_get_elapsed}};
  \code{\link{Stopwatch_reset}};
  \code{\link{Stopwatch_show}};
  \code{\link{Stopwatch_start}};
  \code{\link{Stopwatch_stop}};
  \code{\link{Stopwatch_to_string}}
}

