% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stopwatch.R
\name{Stopwatch_elapsed}
\alias{Stopwatch_elapsed}
\title{Report the current elapsed time without stopping the clock}
\arguments{
\item{...}{further arguments for the \code{get_elapsed} method}
}
\value{
the POSIXct stop time (invisibly)
}
\description{
Report the current elapsed time without stopping the clock
}
\seealso{
Other Stopwatch: \code{\link{Stopwatch_add_elapsed}},
  \code{\link{Stopwatch_current_elapsed}},
  \code{\link{Stopwatch_get_elapsed}},
  \code{\link{Stopwatch_reset}},
  \code{\link{Stopwatch_show}},
  \code{\link{Stopwatch_start}},
  \code{\link{Stopwatch_stop}},
  \code{\link{Stopwatch_to_string}}
}

