% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stopwatch.R
\name{Stopwatch_get_elapsed}
\alias{Stopwatch_get_elapsed}
\title{Get the elapsed time, possibly accumulating.  If the watch is running then 
it is current less start time.  Users should not use this method directly but
use the \code{elapsed} method instead.}
\value{
the difftime (invisibly)
}
\description{
Get the elapsed time, possibly accumulating.  If the watch is running then 
it is current less start time.  Users should not use this method directly but
use the \code{elapsed} method instead.
}
\seealso{
Other Stopwatch: \code{\link{Stopwatch_add_elapsed}},
  \code{\link{Stopwatch_current_elapsed}},
  \code{\link{Stopwatch_elapsed}},
  \code{\link{Stopwatch_reset}},
  \code{\link{Stopwatch_show}},
  \code{\link{Stopwatch_start}},
  \code{\link{Stopwatch_stop}},
  \code{\link{Stopwatch_to_string}}
}
\concept{Stopwatch}
