% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Stopwatch.R
\name{Stopwatch_reset}
\alias{Stopwatch_reset}
\title{Reset the stopwatch, eliminate acculumated elapsed time intervals. Also
if the watch was running, stop it.}
\description{
Reset the stopwatch, eliminate acculumated elapsed time intervals. Also
if the watch was running, stop it.
}
\seealso{
Other Stopwatch: \code{\link{Stopwatch_add_elapsed}};
  \code{\link{Stopwatch_current_elapsed}};
  \code{\link{Stopwatch_elapsed}};
  \code{\link{Stopwatch_get_elapsed}};
  \code{\link{Stopwatch_show}};
  \code{\link{Stopwatch_start}};
  \code{\link{Stopwatch_stop}};
  \code{\link{Stopwatch_to_string}}
}

