% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Stopwatch.R
\name{Stopwatch_show}
\alias{Stopwatch_show}
\title{Show the elapsed time}
\arguments{
\item{...}{futher arguments}
}
\description{
Show the elapsed time
}
\seealso{
Other Stopwatch: \code{\link{Stopwatch_add_elapsed}};
  \code{\link{Stopwatch_current_elapsed}};
  \code{\link{Stopwatch_elapsed}};
  \code{\link{Stopwatch_get_elapsed}};
  \code{\link{Stopwatch_reset}};
  \code{\link{Stopwatch_start}};
  \code{\link{Stopwatch_stop}};
  \code{\link{Stopwatch_to_string}}
}

