% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Stopwatch.R
\name{Stopwatch_start}
\alias{Stopwatch_start}
\title{Start the timer, resets the start time to the current time}
\description{
Start the timer, resets the start time to the current time
}
\seealso{
Other Stopwatch: \code{\link{Stopwatch_add_elapsed}};
  \code{\link{Stopwatch_current_elapsed}};
  \code{\link{Stopwatch_elapsed}};
  \code{\link{Stopwatch_get_elapsed}};
  \code{\link{Stopwatch_reset}};
  \code{\link{Stopwatch_show}};
  \code{\link{Stopwatch_stop}};
  \code{\link{Stopwatch_to_string}}
}

