% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stopwatch.R
\name{Stopwatch_stop}
\alias{Stopwatch_stop}
\title{Stop the timer}
\arguments{
\item{accumulate}{logical, if TRUE then add the elapsed time to the
previous elapsed time}

\item{...}{futher arguments for \code{add_elpased} and \code{get_elapsed} method}
}
\value{
the POSIXct elapsed time
}
\description{
Stop the timer
}
\seealso{
Other Stopwatch: \code{\link{Stopwatch_add_elapsed}},
  \code{\link{Stopwatch_current_elapsed}},
  \code{\link{Stopwatch_elapsed}},
  \code{\link{Stopwatch_get_elapsed}},
  \code{\link{Stopwatch_reset}},
  \code{\link{Stopwatch_show}},
  \code{\link{Stopwatch_start}},
  \code{\link{Stopwatch_to_string}}
}
