% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stopwatch.R
\name{Stopwatch_to_string}
\alias{Stopwatch_to_string}
\title{Return a formatted elapsed time string}
\arguments{
\item{...}{futher arguments for \code{\link{format}} function accepting \code{\link{difftime}}}
}
\description{
Return a formatted elapsed time string
}
\seealso{
Other Stopwatch: \code{\link{Stopwatch_add_elapsed}},
  \code{\link{Stopwatch_current_elapsed}},
  \code{\link{Stopwatch_elapsed}},
  \code{\link{Stopwatch_get_elapsed}},
  \code{\link{Stopwatch_reset}},
  \code{\link{Stopwatch_show}},
  \code{\link{Stopwatch_start}},
  \code{\link{Stopwatch_stop}}
}

