% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{str_trim}
\alias{str_trim}
\title{Trim leading spaces, trailing spaces or both.}
\usage{
str_trim(x, where = c("leading", "trailing", "both")[3], replace_with = "")
}
\arguments{
\item{x}{character vector}

\item{where}{character of either 'leading', 'trailing', 'both' (default)}

\item{replace_with}{character, by default ""}
}
\value{
trimmed character vector of same length as the input
}
\description{
Trim leading spaces, trailing spaces or both.
}

