% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{strip_extension}
\alias{strip_extension}
\title{Strip one or more extensions off one or more filenames}
\usage{
strip_extension(x = c("/A/A.txt", "/B/b/B.csv.txt", "/C/c.C/C.foo"),
  greedy = FALSE)
}
\arguments{
\item{x}{character vector of one or more filenames}

\item{greedy}{logical if TRUE remove all extensions}
}
\value{
filenames without extensions
}
\description{
Greedy behavior is optional.  Given filename \code{/B/b/B.csv.txt}
    greedy behavior yields \code{/B/b/B} while lazy behavior yields \code{/B/b/B.csv}
}
\seealso{
\url{https://stackoverflow.com/questions/2301285/what-do-lazy-and-greedy-mean-in-the-context-of-regular-expressions}
}
