% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{system_memory}
\alias{system_memory}
\title{Retrieves information on the system memory resources on platforms where the
`free` command exists.}
\usage{
system_memory(extra = "")
}
\arguments{
\item{extra}{character extra arguments for the free command}
}
\value{
data.frame or NULL
}
\description{
Retrieves information on the system memory resources on platforms where the
`free` command exists.
}
