% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/compressed.R
\name{test_compressed}
\alias{test_compressed}
\title{Test if a given file is compressed by zip or tgz, gzip.}
\usage{
test_compressed(filename)
}
\arguments{
\item{filename}{the fully qualified name of the file to test}
}
\value{
a character vector of 'zip', 'gzip' or "tgz", "bzip2",  "unknown"
}
\description{
Test if a given file is compressed by zip or tgz, gzip.
}
\seealso{
Other COMPRESSED: \code{\link{is_bzip2}};
  \code{\link{is_gzip}}; \code{\link{is_tgz}};
  \code{\link{is_zip}}; \code{\link{pack_directory}};
  \code{\link{un_bzip2}}; \code{\link{un_gz}};
  \code{\link{un_tgz}}; \code{\link{un_zip}};
  \code{\link{unpack_archive}}
}

