% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Timekeeper.R
\name{Timekeeper}
\alias{Timekeeper}
\title{Create a Timekeeper which holds one or more stopwatches on its fob.}
\usage{
Timekeeper(name = 1, start_watch = TRUE)
}
\arguments{
\item{name}{if start_watch is TRUE, then a name must be provided}

\item{start_watch}{logical, if TRUE start a new Stopwatch}
}
\value{
an instance of TimekeeperRefClass
}
\description{
Create a Timekeeper which holds one or more stopwatches on its fob.
}
\examples{
\dontrun{
   # create a container class, starting a stopwatch right away
   timer <- Timekeeper(start_watch = TRUE, name = 'master')
   Sys.sleep(3)
   # add another stopwatch
   timer$start("another")
   # let's see what we have so far
   timer
   timer$stop("another")
   timer$elapse("master") }
}
\seealso{
Other Stopwatch Timekeeper: \code{\link{Stopwatch}},
  \code{\link{Timekeeper_elapsed}},
  \code{\link{Timekeeper_has_watch}},
  \code{\link{Timekeeper_reset}},
  \code{\link{Timekeeper_start}},
  \code{\link{Timekeeper_stop}}
}

