% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Timekeeper.R
\name{Timekeeper_elapsed}
\alias{Timekeeper_elapsed}
\title{Determine elapsed time for a stopwatch}
\arguments{
\item{name}{either a name or index number (1,2,3,...) for the stopwatch}

\item{to_string}{logical, if TRUE then call the Stopwatch's to_string method}

\item{...}{further arguments for the \code{to_string} method}
}
\description{
Determine elapsed time for a stopwatch
}
\seealso{
Other Stopwatch Timekeeper: \code{\link{Stopwatch}},
  \code{\link{Timekeeper_has_watch}},
  \code{\link{Timekeeper_reset}},
  \code{\link{Timekeeper_start}},
  \code{\link{Timekeeper_stop}}, \code{\link{Timekeeper}}
}
\concept{Stopwatch Timekeeper}
