% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Timekeeper.R
\name{Timekeeper_has_watch}
\alias{Timekeeper_has_watch}
\title{Determine if a specified stopwatch exists}
\arguments{
\item{name}{either a name or index number (1,2,3,...) for the stopwatch}
}
\value{
logical TRUE if the specified name points to a Stopwatch
}
\description{
Determine if a specified stopwatch exists
}
\seealso{
Other Stopwatch Timekeeper: \code{\link{Stopwatch}},
  \code{\link{Timekeeper_elapsed}},
  \code{\link{Timekeeper_reset}},
  \code{\link{Timekeeper_start}},
  \code{\link{Timekeeper_stop}}, \code{\link{Timekeeper}}
}
\concept{Stopwatch Timekeeper}
