% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Timekeeper.R
\name{Timekeeper_reset}
\alias{Timekeeper_reset}
\title{Reset a stopwatch}
\arguments{
\item{name}{either a name or index number (1,2,3,...) for the stopwatch}
}
\description{
Reset a stopwatch
}
\seealso{
Other Stopwatch Timekeeper: \code{\link{Stopwatch}},
  \code{\link{Timekeeper_elapsed}},
  \code{\link{Timekeeper_has_watch}},
  \code{\link{Timekeeper_start}},
  \code{\link{Timekeeper_stop}}, \code{\link{Timekeeper}}
}

