% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Timekeeper.R
\name{Timekeeper_start}
\alias{Timekeeper_start}
\title{Start a stopwatch}
\arguments{
\item{name}{either a name or index number (1,2,3,...) for the stopwatch}
}
\description{
Start a stopwatch
}
\seealso{
Other Stopwatch Timekeeper: \code{\link{Stopwatch}};
  \code{\link{Timekeeper_elapsed}};
  \code{\link{Timekeeper_has_watch}};
  \code{\link{Timekeeper_reset}};
  \code{\link{Timekeeper_stop}}; \code{\link{Timekeeper}}
}

