% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Timekeeper.R
\name{Timekeeper_stop}
\alias{Timekeeper_stop}
\title{Stop a stopwatch}
\arguments{
\item{name}{either a name or index number (1,2,3,...) for the stopwatch}

\item{...}{further arguments for StopWatch (espcially interval)}
}
\description{
Stop a stopwatch
}
\seealso{
Other Stopwatch Timekeeper: \code{\link{Stopwatch}},
  \code{\link{Timekeeper_elapsed}},
  \code{\link{Timekeeper_has_watch}},
  \code{\link{Timekeeper_reset}},
  \code{\link{Timekeeper_start}}, \code{\link{Timekeeper}}
}
\concept{Stopwatch Timekeeper}
