% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Timekeeper.R
\docType{class}
\name{TimekeeperRefClass-class}
\alias{TimekeeperRefClass}
\alias{TimekeeperRefClass-class}
\title{A class for storing one or more StopwatchRefClass objects}
\description{
A class for storing one or more StopwatchRefClass objects
}
\section{Fields}{

\describe{
\item{\code{fob}}{a lost of StopwatchRefClass objects}
}}

