% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compressed.R
\name{un_bzip2}
\alias{un_bzip2}
\title{Unpack a tarred and bzip2-ped archive.}
\usage{
un_bzip2(filename, extra = "-xf")
}
\arguments{
\item{filename}{the fully qualified name of the file to unpack}

\item{extra}{extra arguments for the system's \code{tar} command}
}
\value{
a character vector of the uncompressed filenames including relative paths
}
\description{
Unpack a tarred and bzip2-ped archive.
}
\seealso{
Other COMPRESSED: \code{\link{is_bzip2}},
  \code{\link{is_gzip}}, \code{\link{is_tgz}},
  \code{\link{is_zip}}, \code{\link{pack_directory}},
  \code{\link{test_compressed}}, \code{\link{un_gz}},
  \code{\link{un_tgz}}, \code{\link{un_zip}},
  \code{\link{unpack_archive}}
}

