% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compressed.R
\name{un_gz}
\alias{un_gz}
\title{Unpack a gzipped archive.}
\usage{
un_gz(filename, out = NA, extra = "-d")
}
\arguments{
\item{filename}{the fully qualified name of the file to unpack}

\item{out}{the name of the outfile to create, if NA then it is whatever gets unpacked.
If you specify \code{out} be sure that \code{extra} gets at least '-cd'}

\item{extra}{extra arguments for the system's \code{gzip} command}
}
\value{
a character vector of the uncompressed filenames including relative paths
}
\description{
Unpack a gzipped archive.
}
\seealso{
Other COMPRESSED: \code{\link{is_bzip2}},
  \code{\link{is_gzip}}, \code{\link{is_tgz}},
  \code{\link{is_zip}}, \code{\link{pack_directory}},
  \code{\link{test_compressed}}, \code{\link{un_bzip2}},
  \code{\link{un_tgz}}, \code{\link{un_zip}},
  \code{\link{unpack_archive}}
}
\concept{COMPRESSED}
