% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/compressed.R
\name{un_zip}
\alias{un_zip}
\title{Unzip a zipped archive.}
\usage{
un_zip(filename, extra = "-x *MACOSX*/* *DS_Store*")
}
\arguments{
\item{filename}{the fully qualified name of the file to unzip}

\item{extra}{extra arguments for the system's \code{unzip} command}
}
\value{
a character vector of the uncompressed filenames including relative paths
}
\description{
Unzipping a file can be a bit tricky if the file was zipped on a Mac OSX
which may include MACOSX/ and DS_STORE stuff we don't usually want.
This function tries to clean up the output.
}
\seealso{
Other COMPRESSED: \code{\link{is_bzip2}};
  \code{\link{is_gzip}}; \code{\link{is_tgz}};
  \code{\link{is_zip}}; \code{\link{pack_directory}};
  \code{\link{test_compressed}}; \code{\link{un_bzip2}};
  \code{\link{un_gz}}; \code{\link{un_tgz}};
  \code{\link{unpack_archive}}
}

