% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compressed.R
\name{unpack_archive}
\alias{unpack_archive}
\title{Unpack a compressed archive: zip, gzip, and gzipped tarballs.}
\usage{
unpack_archive(filename, ...)
}
\arguments{
\item{filename}{the fully qualified name(s) of the file to unpack}

\item{...}{extra arguments for the system's  decompression command (unzip, gzip -d, tar -x)}
}
\value{
a character vector of the uncompressed filenames including relative paths
   or a list of such if multiple filenames are provided
}
\description{
This is a loose routine to unpack and list the contents of a file, possibely
a compressed file. If a directory is provided then return a list of the contents,  if an
uncompressed file (or some other thing) is provided then return just that.
}
\details{
See \code{\link{un_zip}}, \code{\link{un_gz}} and \code{\link{un_tgz}}
}
\seealso{
Other COMPRESSED: \code{\link{is_bzip2}};
  \code{\link{is_gzip}}; \code{\link{is_tgz}};
  \code{\link{is_zip}}; \code{\link{pack_directory}};
  \code{\link{test_compressed}}; \code{\link{un_bzip2}};
  \code{\link{un_gz}}; \code{\link{un_tgz}};
  \code{\link{un_zip}}
}

