% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mail.R
\name{has_mutt}
\alias{has_mutt}
\title{Test if mutt is installed on the host platform
See \url{http://www.mutt.org/}}
\usage{
has_mutt()
}
\value{
logical, TRUE if present
}
\description{
Test if mutt is installed on the host platform
See \url{http://www.mutt.org/}
}
\seealso{
Other mail: \code{\link{has_nail}}; \code{\link{muttmail}};
  \code{\link{nailmail}}; \code{\link{sendmail}}
}

