% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mail.R
\name{sendmail}
\alias{sendmail}
\title{Send an email from R using the underlying mail services.}
\usage{
sendmail(..., what = c("nail", "mutt")[1])
}
\arguments{
\item{...}{arguments for the mailer of choice. Currently, only mutt is supported.}

\item{what}{the name of the mail engine to use, currently 'nail' and 'mutt' are supported}
}
\value{
value returned by the system-level mail program, non-zero for failure.
}
\description{
A simple wrapper around the platforms email engine.  Currently \code{mutt} and
\code{nail} are supported, but the host must have these installed.
}
\seealso{
Other mail: \code{\link{has_mutt}}, \code{\link{has_nail}},
  \code{\link{muttmail}}, \code{\link{nailmail}}
}

